pyproject.toml
[project]
name = "rag_with_crewai_flow"
version = "0.1.0"
description = "rag-with-crewai-flow using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.9.3",
    "google-auth>=2.48.0",
    "google-auth-oauthlib>=1.3.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.36.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.10.1",
    "vertexai>=1.71.1",
]

[project.scripts]
kickoff = "rag_with_crewai_flow.main:kickoff"
run_crew = "rag_with_crewai_flow.main:kickoff"
plot = "rag_with_crewai_flow.main:plot"
run_with_trigger = "rag_with_crewai_flow.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
